Very basic start (needs editing)
This is a very basic start, it should be accepted as a valid child
theme, but it doesn't have a lot of the basic setup things done (like
defining the correct parent theme, the name of the theme, or proper
enqueing of the CSS, and a lot of other things).

I will be updating those things right away.